redactedbuilder/requirements.txt
solders>=0.21.0
solana>=0.34.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
